requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.13
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
